Refactor to introduce centralised knowledge of inverter registers and capabilities
The motivation is twofold:

1. Generating an optimal sequence of register reads for a particular
   interver. This means knowing all of the address of all registers on
   a particular inverter
2. Writing force charge windows is probably going to need some centralised
   knowledge about an inverter (such as which registers need to be written
   in one block), and currently the knowledge about an inverter is spread
   across the codebase

The central points are inverter_connection_types, a list of ways to connect
to inverters, and and inverter_profiles, a list of all inverter models
which are supported, along with inverter-model-specific metadata, and the
list of entities which are supported by that inverter model.

All parts of the codebase which need to deal with the differences between
inverter models and connection types now go through these central points.