Dramatically speed up TCP reads
So, I'm using the W610, which as you've observed is quite slow to poll.

However, I noticed in the log that some reads were happening pretty
quickly, which seemed odd. On a hunch I checked the pymodbus source,
and lo and behold it doesn't disable Nagle's algorithm. This is despite
the modbus documentation recommending that it be disabled.

See Section 4.3.2:
https://www.modbus.org/docs/Modbus_Messaging_Implementation_Guide_V1_0b.pdf

So, most of the delay was due to the TCP stack hanging around to see
whether we were going to be sending any more data.

Update ModbusClient to disable Nagle's algorithm.