Remove the need for separate lists of entity types
I've been wanting to do this for a while. Instead of having to keep
separate lists of the different types of entities, which is only going
to grow when we start adding integrals and other things, teach each
EntityDescription how to create an instance of its corresponding Entity.

This means we can mash everything into one big list of entity
descriptions, and when HA asks for all binary sensors we find the ones
which will create binarys sensors, and ask them to create those.