Handle polls which are aborted partway through
Previously, if a later read failed, we would records new values for
earlier reads but wouldn't notify the sensors which depended on those
reads.

This change ensures that we perform all of the reads first, and only if
all succeed do we write those to _data and notify the sensors.

Relates-to: #122