Remove ModbusClient.connect
It turns out that pymodbus automatically connects the first time that you
try and read/write something anyway. I think that the initial first
auto_connect connection was messing up my W610 as well, as I often got
a failed read on the first read attempt, and would have to wait for the
automatic reconnection to happen. Removing the auto_connect seems to fix
this.