Fetch entity addresses from the entity itself rathar than its EntityFactory
Previously, the EntityFactory knew all of the addressses that the entity
it described would rely on. This was passed to the
InverterModelConnectionTypeProfile, which compiled the list of all
addresses the controller cared about.

Move to a model where the entity itself specifies its addresses. Instead
of the addresses being passed through the
InverterModelConnectionTypeProfile, instead we now record an entity/s
addresses when it registers with the controller (and remove them when it
unregisters).

This doesn't change much right now, but it has two important consequences:
1. An entity's addresses can now be set when it is instantiated, rather
   than being fixed on its description. This lets an entity have different
   addresses depending on the inverter model it's placed on.
2. We now have the option of unregistering an entity when it's disabled, so
   it won't get updates.