Combine all entities into a single list
The idea here is that we have a single list of entities descriptions. Each
descriptions says the types of inverters and connection types it supports,
and for each combination the address(es) of the register(s) it relys on.

When we come to actually create the entities, we filter to the ones which
support our inverter / connection type combination, and use the address
from that combination.

This caught a bug in the LAN sensors: the inverter/ambient sensors were
swapped for AUX, but not for LAN.