Support pymodbus 3.4.1
HA has upgraded the version it uses. Since we specify a lower version, if
there are any other integrations which use the higher version we'll race
to see which gets installed first, so there's a chance that we'll end
up with 3.4.1.

This matters, because 3.4.1 doesn't have automatic (re)connections.

Fix this. Also relax our range to >=3.1.3, so that if HA upgrades its
version again we'll deterministically get the newer version, and if there's
a problem we'll know about it, rather than being in a sometimes-working
state.

Fixes: #393