Don't assume that entity_ids are what we expect
For things which rely on the entity ids of other entities (integration
sensors, lambda sensors, charge period config for the card), don't assume
that the entity id is as we set it. Instead, use the unique id to look up
the entity id in the entity registry.

I've been meaning to do this for a while, and it works around an issue in
1.7.1-1.7.2 where we used the unique_id_prefix instead of the
entity_id_prefix for entity ids, and also didn't set the unique_id_prefix
during setup. This means that all entities don't have the entity id prefix.