Add ability to change Inv based on selected inverter manager version
We don't (yet?) have the option to automatically switch based on the
version reported by the inverter -- that's something I've tried and failed
to do a couple of times. But it adds an option to the settings to select
your manager version, which at least means that people have the option
of not being broken.