Support ConnectionType.LAN implicitly for all inverter models
It looks like we're not really aware of which inverters support LAN and
which don't, and there's nothing to stop a user from selecting the wrong
one during setup anyway.

If we're looking for a connection type of LAN on a profile, and we don't
find it, then look for AUX as a backup.

Fixes: #817